Move test_report_size to TestQuantity.test_size
